Added one more perspective to the correlation scan.

+++ src/trainer.hpp
@@ -35,6 +35,6 @@ private:
 	void saveBrains();
 
 public:
-	void resume(std::string session, size_t round);
+	void resume(std::string session, int round);
 	void train();
 };
